+++ boris-client/src/Boris/Client/Log.hs
@@ -6,10 +6,8 @@ module Boris.Client.Log (
 
 import           Boris.Client.Http (BorisHttpClientError (..))
 import qualified Boris.Client.Http as H
-import           Boris.Core.Data
-import           Boris.Core.Data.Log
 import           Boris.Core.Data.Build
-import           Boris.Core.Data.Project
+import           Boris.Core.Data.Log
 import           Boris.Representation.ApiV1
 
 import           P
