+++ boris-core/src/Boris/Core/Data/Agent.hs
@@ -2,6 +2,7 @@
 module Boris.Core.Data.Agent (
     Agent (..)
   , AgentId (..)
+  , QueueSize (..)
   ) where
 
 import           P
@@ -15,4 +16,9 @@ data Agent =
 newtype AgentId =
   AgentId {
       agentId :: Int64
+    } deriving (Eq, Ord, Show)
+
+newtype QueueSize =
+  QueueSize {
+      getQueueSize :: Int
     } deriving (Eq, Ord, Show)
