+++ boris-core/test/Test/Boris/Core/Arbitrary.hs
@@ -7,7 +7,6 @@ import qualified Data.Text as T
 
 import           Disorder.Corpus
 
-import           Boris.Core.Data
 import           Boris.Core.Data.Build
 import           Boris.Core.Data.Configuration
 import           Boris.Core.Data.Project
