+++ boris-core/test/test.hs
@@ -1,4 +1,4 @@
-import qualified Test.Boris.Core.Data
+import qualified Test.Boris.Core.Data.Build
 import qualified Test.Boris.Core.Serial.Command
 import qualified Test.Boris.Core.Serial.Ref
 
@@ -8,7 +8,7 @@ import           Disorder.Core.Main
 main :: IO ()
 main =
   disorderMain [
-      Test.Boris.Core.Data.tests
+      Test.Boris.Core.Data.Build.tests
     , Test.Boris.Core.Serial.Ref.tests
     , Test.Boris.Core.Serial.Command.tests
     ]
