+++ boris-http/src/Boris/Http/View.hs
@@ -22,7 +22,6 @@ import           BMX (BMXError, renderBMXError)
 import           BMX (BMXState, Template, partialFromTemplate, renderPage, renderTemplate, templateFile)
 import           BMX (BMXValue (..), defaultState, usingContext, usingPartials)
 
-import           Boris.Core.Data
 import           Boris.Core.Data.Build
 import           Boris.Core.Data.Project
 import           Boris.Http.Data
