+++ boris-representation/src/Boris/Representation/ApiV1.hs
@@ -22,7 +22,7 @@ module Boris.Representation.ApiV1 (
   , CreateProject (..)
   ) where
 
-import           Boris.Core.Data
+import           Boris.Core.Data.Agent
 import           Boris.Core.Data.Build
 import           Boris.Core.Data.Instance
 import           Boris.Core.Data.Log
